% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_ggplot}
\alias{assert_ggplot}
\title{Assert that an object is a ggplot}
\usage{
assert_ggplot(p, arg_name = deparse(substitute(p)))
}
\arguments{
\item{p}{An object to check.}

\item{arg_name}{Character string used in the error message to name the
argument. Defaults to the deparsed expression of \code{p}.}
}
\value{
\code{p} invisibly, if it passes the check.
}
\description{
Stops with an informative error if \code{p} is not a ggplot object.
}
\examples{
p <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy))
assert_ggplot(p)
}
