% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspec_result.R
\name{combine_results}
\alias{combine_results}
\title{Combine multiple \code{ggspec_result} objects into one}
\usage{
combine_results(results)
}
\arguments{
\item{results}{A list of \code{ggspec_result} objects.}
}
\value{
A single \code{ggspec_result}.
}
\description{
The combined result passes only if all sub-results pass. The message
summarises how many checks passed.
}
\keyword{internal}
