% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{equiv_coord}
\alias{equiv_coord}
\title{Compare coordinate systems of two ggplot objects}
\usage{
equiv_coord(p1, p2)
}
\arguments{
\item{p1}{Reference ggplot or \code{ggspec_canon} object.}

\item{p2}{Observed ggplot or \code{ggspec_canon} object.}
}
\value{
A \code{ggspec_result}.
}
\description{
Compare coordinate systems of two ggplot objects
}
\examples{
p1 <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
  ggplot2::geom_point()
p2 <- p1 + ggplot2::coord_flip()
equiv_coord(p1, p2)
}
