% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{equiv_facets}
\alias{equiv_facets}
\title{Compare facet specification of two ggplot objects}
\usage{
equiv_facets(p1, p2)
}
\arguments{
\item{p1}{Reference ggplot or \code{ggspec_canon} object.}

\item{p2}{Observed ggplot or \code{ggspec_canon} object.}
}
\value{
A \code{ggspec_result}.
}
\description{
Compare facet specification of two ggplot objects
}
\examples{
p1 <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
  ggplot2::geom_point() + ggplot2::facet_wrap(~class)
p2 <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
  ggplot2::geom_point() + ggplot2::facet_wrap(~drv)
equiv_facets(p1, p2)
}
