% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{equiv_labels}
\alias{equiv_labels}
\title{Compare labels of two ggplot objects}
\usage{
equiv_labels(p1, p2, aesthetics = NULL)
}
\arguments{
\item{p1}{Reference ggplot or \code{ggspec_canon} object.}

\item{p2}{Observed ggplot or \code{ggspec_canon} object.}

\item{aesthetics}{Character vector of label names to compare. \code{NULL}
(default) compares all labels present in \code{p1}.}
}
\value{
A \code{ggspec_result}.
}
\description{
Compare labels of two ggplot objects
}
\examples{
p1 <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
  ggplot2::geom_point() + ggplot2::labs(title = "My plot")
p2 <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
  ggplot2::geom_point() + ggplot2::labs(title = "Different title")
equiv_labels(p1, p2)
equiv_labels(p1, p2, aesthetics = "x")  # passes (x labels same)
}
