% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{equiv_params}
\alias{equiv_params}
\title{Compare parameters of a specific layer in two ggplot objects}
\usage{
equiv_params(p1, p2, layer = 1L, params = NULL)
}
\arguments{
\item{p1}{Reference ggplot or \code{ggspec_canon} object.}

\item{p2}{Observed ggplot or \code{ggspec_canon} object.}

\item{layer}{Integer: which layer index to compare (1-based). Compared by
position.}

\item{params}{Character vector of parameter names to check. \code{NULL}
(default) checks all parameters present in \code{p1}'s layer.}
}
\value{
A \code{ggspec_result}.
}
\description{
Compare parameters of a specific layer in two ggplot objects
}
\examples{
p1 <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
  ggplot2::geom_smooth(method = "lm", se = FALSE)
p2 <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
  ggplot2::geom_smooth(method = "loess", se = TRUE)
equiv_params(p1, p2, layer = 1L, params = c("se"))
}
