% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{equiv_scales}
\alias{equiv_scales}
\title{Compare scales of two ggplot objects}
\usage{
equiv_scales(p1, p2, aesthetics = NULL)
}
\arguments{
\item{p1}{Reference ggplot or \code{ggspec_canon} object.}

\item{p2}{Observed ggplot or \code{ggspec_canon} object.}

\item{aesthetics}{Character vector of aesthetics to compare. \code{NULL}
(default) compares all scales present in \code{p1}.}
}
\value{
A \code{ggspec_result}.
}
\description{
Compare scales of two ggplot objects
}
\examples{
p1 <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
  ggplot2::geom_point() + ggplot2::scale_x_log10()
p2 <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
  ggplot2::geom_point()
equiv_scales(p1, p2)
}
