% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_ggplot}
\alias{is_ggplot}
\title{Test whether an object is a ggplot}
\usage{
is_ggplot(p)
}
\arguments{
\item{p}{An object to test.}
}
\value{
\code{TRUE} if \code{p} inherits from \code{"gg"}, \code{FALSE} otherwise.
}
\description{
Test whether an object is a ggplot
}
\examples{
p <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy))
is_ggplot(p)
is_ggplot(list())
}
