% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_layers}
\alias{n_layers}
\title{Count the number of layers in a ggplot}
\usage{
n_layers(p)
}
\arguments{
\item{p}{A ggplot object.}
}
\value{
A non-negative integer.
}
\description{
Count the number of layers in a ggplot
}
\examples{
p <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
  ggplot2::geom_point() +
  ggplot2::geom_smooth()
n_layers(p)
}
